change review rate Integer to Flout

+++ src/main/java/com/filot/filotshop/review/entity/ReviewForm.java
@@ -8,6 +8,5 @@ import java.time.LocalDateTime;
 public class ReviewForm {
     private String content;
     private String title;
-    private Integer rate;
-
+    private Float rate;
 }
